Complete of rejig

+++ stackstrap/repository.py
@@ -16,7 +16,7 @@ class Repository(object):
         self.log = logging.getLogger("repository")
         self.url = url
 
-    def archive_to(self, git_ref, destination, *archive_args):
+    def archive(self, ref, destination, *archive_args):
         """
         Archive the specified GIT ref to the specified destination
 
@@ -24,38 +24,32 @@ class Repository(object):
         add extra flags for `git archive` should you desire.
         """
         tmp_dir = tempfile.mkdtemp()
-
-        # create our git interface via the sh module
-        # the loose_ssh.sh script is a simple wrapper that sets:
-        #     StrictHostKeyChecking=no
-        #
-        # see:
-        # github.com/openops/stackstrap/blob/master/scripts/loose_ssh.sh
-        self.git = sh.git.bake(_cwd=tmp_dir, _env={
-            'GIT_SSH': "loose_ssh.sh"
-        })
-
-        self.log.debug("Creating a new copy of the repository in our \
-                        cache (%s)..." % self.path)
-        self.git('clone', '--recursive', self.url, tmp_dir)
-
-        self.log.debug("Archiving '{ref}' to {destination}".format(
-            ref=git_ref,
-            destination=destination
-        ))
-
         tar_file = None
-        settings.mkdir_p(destination)
 
         try:
+            # create our git interface via the sh module
+            # see:
+            # github.com/openops/stackstrap/blob/master/scripts/loose_ssh.sh
+            git = sh.git.bake(_cwd=tmp_dir, _env={
+                'GIT_SSH': "loose_ssh.sh"
+            })
+
+            self.log.debug("Cloning %s to %s" % (self.url, tmp_dir))
+            git('clone', '--recursive', self.url, tmp_dir)
+
+            self.log.debug("Archiving '{ref}' to {destination}".format(
+                ref=ref,
+                destination=destination
+            ))
+
+            settings.mkdir_p(destination)
+
             (fd, tar_file) = tempfile.mkstemp()
-            self.git.archive(git_ref,
-                             *archive_args, _out=tar_file)
-            sh.tar("xf", tar_file, _cwd=destination)
+            git.archive('origin/%s' % ref, *archive_args, _out=tar_file)
+            sh.tar("xpf", tar_file, _cwd=destination)
         finally:
             if tar_file:
                 os.remove(tar_file)
-            shutil.rmtree(tmp_dir)
 
-    def cache_path(self, *parts):
-        return os.path.join(self.path, *parts)
+            if os.path.isdir(tmp_dir):
+                shutil.rmtree(tmp_dir)
